% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevomod-spatial-package.R
\docType{package}
\name{cevomod.spatial-package}
\alias{cevomod.spatial}
\alias{cevomod.spatial-package}
\title{cevomod.spatial: cevomod package extension for spatial data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Paweł Kuś \email{kpawel2210@gmail.com}

}
\keyword{internal}
