% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_space}
\alias{plot_space}
\alias{plot_empty_space}
\title{Space plots}
\usage{
plot_space(object, ...)

plot_empty_space(object, fill = "gray90", ...)
}
\arguments{
\item{object}{cevodata object}

\item{...}{other parameters passed to geom_tile()}

\item{fill}{fill}
}
\description{
Space plots
}
